moved cleanup method to common
